+++ core/interactions/deps.ts
@@ -1,0 +1,1 @@
+export * from "./https://unpkg.com/@evan/wasm@0.0.65/target/ed25519/deno.js";