+++ core/util/src/utf8_codec.ts
@@ -1,7 +1,7 @@
 // deno-lint-ignore-file no-explicit-any
 
 export const decode = (globalThis as any).Deno?.core.decode ??
-  TextDecoder.prototype.decode.bind(new TextDecoder());
+  ((d) => (input: BufferSource) => d.decode(input))(new TextDecoder());
 
-export const encode = (globalThis as any).Deno?.core.decode ??
-  TextEncoder.prototype.encode.bind(new TextEncoder());
+export const encode = (globalThis as any).Deno?.core.encode ??
+  ((e) => (input: string) => e.encode(input))(new TextEncoder());